Moved train/val scripts to scripts/
